requirements.txt
aiopg==1.4.0
alembic==1.11.1
anyio==3.7.0
async-timeout==4.0.2
asyncpg==0.27.0
bcrypt==4.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
deprecation==2.1.0
dnspython==2.3.0
docker==6.1.3
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
fastapi==0.97.0
flake8==6.1.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
iniconfig==2.0.0
Jinja2==3.1.2
load-dotenv==0.1.0
Mako==1.2.4
MarkupSafe==2.1.3
mccabe==0.7.0
packaging==23.1
passlib==1.7.4
pluggy==1.2.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
pyasn1==0.5.0
pycodestyle==2.11.0
pydantic==1.10.9
pyflakes==3.1.0
pytest==7.4.0
pytest-asyncio==0.21.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.16
starlette==0.27.0
testcontainers==3.7.1
tomli==2.0.1
typing-extensions==4.6.3
urllib3==2.0.4
uvicorn==0.22.0
websocket-client==1.6.1
websockets==11.0.3
wrapt==1.15.0
zipp==3.15.0
